Modified New Dataset Creation
Now each song only has one genre (some had multiple), and the new set will consist of 2000 songs of each of the categories that are listed in the cleaner. It will search until it has found 2000 songs of each but only write the song to the csv if 2000 hasn't been reached yet. Can also be changed but looking at the distribution of genres over the entire set the lesser popular genres only have close to 2000 songs.