Final Dataset commit
I guess we could also clean up the other columns like the genre being 'rap' instead of '[rap]'?  could be easier later on but nor sure? 
The script takes 5 hours to complete, therefore I'll just run it and send you the dataset
Next we can analyze the data and create graphs/plots for it and decide on train/test split and that sort of stuff